Добавление моделей для сайта и Fix удаления тегов+категорий

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -99,7 +99,13 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        $category = Category::find($id);
+        if ($category->products->count()) {
+            return redirect()->route('categories.index')->with('error', "Ошибка! У категории есть записи {$category->products->count()}.шт");
+        }
+
         Category::destroy($id);
+
         return redirect()->route('categories.index')->with('success', 'Категория удалена');
     }
 }
