+++ app/Http/Controllers/Admin/TagController.php
@@ -88,7 +88,16 @@ class TagController extends Controller
      */
     public function destroy($id)
     {
+
+//        Tag::destroy($id);
+
+        $tag = Tag::find($id);
+        if ($tag->products->count()) {
+            return redirect()->route('tags.index')->with('error', "Ошибка! У тега есть записи {$tag->products->count()}.шт");
+        }
+
         Tag::destroy($id);
+
         return redirect()->route('tags.index')->with('success', 'Тег удален');
     }
 }
